feat(Pointers): Moved the files to src/Components/Pointers
